requirements.txt
Django==3.2.5
gunicorn==20.1.0
djangorestframework
django-extensions
